# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestão Escolar",
        "description": "API para gerenciamento de sistema escolar.",
        "contact": {
            "email": "suporte@exemplo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/schedule": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "List Schedule with pagination",
                "description": "Get paginated list of Schedules",
                "operationId": "770209026e4bbff78b8f3abeffa313f9",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for filtering schedule",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "sort_column",
                        "in": "query",
                        "description": "Column to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Direction for sorting (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of schedule",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ScheduleResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Create a new Schedule",
                "description": "Create a new Schedule",
                "operationId": "e4e60c75570ca32c9fe2efda7d5d6858",
                "requestBody": {
                    "description": "JSON object containing new Schedule data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScheduleStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Schedule created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScheduleResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/schedule/{id}": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Retrieve details of a Schedule by ID",
                "description": "Show details of a specific Schedule",
                "operationId": "02bf5dcbb2387d48b8b6958559a822b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Schedule ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schedule data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScheduleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Schedule not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Update an existing Schedule",
                "description": "Update an existing Schedule",
                "operationId": "040d99ce796a94d3ab4dd77c5f8e3174",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Schedule ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated Schedule data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScheduleUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Schedule updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScheduleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Schedule not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Delete a Schedule by ID",
                "description": "Delete a specific Schedule",
                "operationId": "34d2b2bb1d32572f1220338b5d2703f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Schedule ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Schedule deleted successfully"
                    },
                    "404": {
                        "description": "Schedule not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/schedules/class/{class_id}": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Get schedules by class ID",
                "description": "Retrieve a list of schedules filtered by class ID",
                "operationId": "0aded81945418f6caaa4a8bba31fa659",
                "parameters": [
                    {
                        "name": "class_id",
                        "in": "path",
                        "description": "ID of the class",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of schedules for the class",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ScheduleResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/schedules/teacher/{teacher_id}": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Get schedules by teacher ID",
                "description": "Retrieve a list of schedules filtered by teacher ID",
                "operationId": "da0531b89df10243d238dc01ef35bcf8",
                "parameters": [
                    {
                        "name": "teacher_id",
                        "in": "path",
                        "description": "ID of the teacher",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of schedules for the teacher",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ScheduleResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/schedules/discipline/{discipline_id}": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Get schedules by discipline ID",
                "description": "Retrieve a list of schedules filtered by discipline ID",
                "operationId": "8f69197f8762875dcab6828a02e7b96e",
                "parameters": [
                    {
                        "name": "discipline_id",
                        "in": "path",
                        "description": "ID of the discipline",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of schedules for the discipline",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ScheduleResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/schedules/day/{day}": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Get schedules by day",
                "description": "Retrieve a list of schedules filtered by a specific day",
                "operationId": "604f1ed71687a10cea4df61fb489c3ee",
                "parameters": [
                    {
                        "name": "day",
                        "in": "path",
                        "description": "Day of the week",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of schedules for the day",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ScheduleResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}